+++ src/main/java/ut/veikotiit/checkers/scorer/DefaultBitBoardScorer.java
@@ -8,7 +8,7 @@ public class DefaultBitBoardScorer implements BitBoardScorer {
     private static BitBoardScorer INSTANCE = new DefaultBitBoardScorer();
 
     @Override
-    public int getScore(BitBoard bitBoard) {
+    public double getScore(BitBoard bitBoard) {
         Color color = bitBoard.getNextColor();
 
         int score = Long.bitCount(bitBoard.getAllBlackPieces()) + Long.bitCount(bitBoard.getBlackKings())
